pyproject.toml
[tool.poetry]
name = "kkutbot"
version = "2.0.0"
description = "끝말잇기 디스코드봇"
authors = ["janu8ry <janu8ry0108@gmail.com>"]
license = "AGPL-3.0"
readme = "README.md"
homepage = "https://kkutbot.github.io"
repository = "https://github.com/janu8ry/kkutbot"

[tool.poetry.dependencies]
python = "^3.9"
"discord.py" = {extras = ["speed"], version = "^2.0.0"}
motor = "^3.0.0"
aiohttp = {extras = ["sppedups"], version = "^3.8.1"}
koreanbots = "^2.1.0"
topggpy = "^1.4.0"
PyYAML = "^6.0"
rich = "^12.4.4"
APScheduler = "^3.9.1"
humanize = "^4.1.0"
jishaku = {extras = ["procinfo"], version = "^2.5.1"}

[tool.poetry.dev-dependencies]
isort = "^5.10.1"
flake8 = "^5.0.4"
black = "^23.3.0"

[tool.black]
line-length = "150"

[tool.isort]
skip_gitignore = true
line_length = "150"
extra_standard_library = ["typing_extensions"]
profile = "black"

[tool.mypy]
strict = false
warn_unused_configs = true
show_error_codes = true
disable_error_code = "misc"
exclude = ".venv"

[[tool.mypy.overrides]]
module = ["discord.*", "motor.*", "apscheduler.*", "yaml.*", "rich.*", "koreanbots.*", "topgg.*", "psutil.*"]
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = ["tools.views"]
ignore_errors = true

[tool.poetry.urls]
Invite = "https://discord.com/api/oauth2/authorize?client_id=703956235900420226&permissions=387136&scope=bot"
discord = "https://discord.gg/z8tRzwf"
koreanbots = "https://koreanbots.dev/bots/703956235900420226"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

